pyproject.toml
[tool.poetry]
name = "nydok"
version = "v0.0"
description = "Documentation and specification testing framework"
authors = ["Nykode Therapeutics <insaid@nykode.com>"]
license = "MIT"
classifiers = []
packages = [
  { include = "nydok" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.3"
PyYAML = "^6.0"
requests = "^2.28.1"
lxml = "^4.9.1"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^3.9.1"
black = "^22.3.0"
pytest-cov = "^2.11.1"
coverage = {extras = ["toml"], version = "^5.5"}
mypy = "^0.910"
isort = "^5.9.3"
mkdocs-material = "^8.5.10"
mkdocs-click = "^0.6.0"
types-PyYAML = "^6.0.7"
pytest-httpserver = "^1.0.5"
types-Markdown = "^3.4.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
nydok = 'nydok.cli:cli'

[tool.poetry.plugins."pytest11"]
nydok = "nydok.plugin.plugin"

[tool.black]
line-length = 100

[tool.coverage.run]
omit = [ "tests/*" ]

[tool.isort]
profile = "black"

# For flake8, see .flake8